+++ services/github/github-release.service.js
@@ -1,28 +1,14 @@
 'use strict'
 
-const Joi = require('@hapi/joi')
 const { addv } = require('../text-formatters')
 const { version: versionColor } = require('../color-formatters')
-const { latest } = require('../version')
 const { GithubAuthV3Service } = require('./github-auth-service')
 const {
   fetchLatestRelease,
-  releaseInfoSchema,
+  queryParamSchema,
 } = require('./github-common-fetch')
-const { documentation, errorMessagesFor } = require('./github-helpers')
-const { NotFound, redirector } = require('..')
-
-const queryParamSchema = Joi.object({
-  include_prereleases: Joi.equal(''),
-  sort: Joi.string()
-    .valid('date', 'semver')
-    .default('date'),
-}).required()
-
-const releaseInfoArraySchema = Joi.alternatives().try(
-  Joi.array().items(releaseInfoSchema),
-  Joi.array().length(0)
-)
+const { documentation } = require('./github-helpers')
+const { redirector } = require('..')
 
 class GithubRelease extends GithubAuthV3Service {
   static get category() {
@@ -100,64 +86,15 @@ class GithubRelease extends GithubAuthV3Service {
     return { message: addv(version), color }
   }
 
-  async fetchReleases({ user, repo }) {
-    return this._requestJson({
-      url: `/repos/${user}/${repo}/releases`,
-      schema: releaseInfoArraySchema,
-      errorMessages: errorMessagesFor('repo not found'),
-    })
-  }
-
-  static getLatestRelease({ releases, sort, includePrereleases }) {
-    if (sort === 'semver') {
-      const latestRelease = latest(
-        releases.map(release => release.tag_name),
-        {
-          pre: includePrereleases,
-        }
-      )
-      const kvpairs = Object.assign(
-        ...releases.map(release => ({ [release.tag_name]: release.prerelease }))
-      )
-      return { tag_name: latestRelease, prerelease: kvpairs[latestRelease] }
-    }
-
-    if (!includePrereleases) {
-      const stableReleases = releases.filter(release => !release.prerelease)
-      if (stableReleases.length > 0) {
-        return stableReleases[0]
-      }
-    }
-
-    return releases[0]
-  }
-
   async handle({ user, repo }, queryParams) {
-    const sort = queryParams.sort
-    const includePrereleases = queryParams.include_prereleases !== undefined
-
-    if (!includePrereleases && sort === 'date') {
-      const latestRelease = await fetchLatestRelease(this, { user, repo })
-      return this.constructor.render({
-        version: latestRelease.tag_name,
-        sort,
-        isPrerelease: latestRelease.prerelease,
-      })
-    }
-
-    const releases = await this.fetchReleases({ user, repo })
-    if (releases.length === 0) {
-      throw new NotFound({ prettyMessage: 'no releases' })
-    }
-    const latestRelease = this.constructor.getLatestRelease({
-      releases,
-      sort,
-      includePrereleases,
-    })
-
+    const latestRelease = await fetchLatestRelease(
+      this,
+      { user, repo },
+      queryParams
+    )
     return this.constructor.render({
       version: latestRelease.tag_name,
-      sort,
+      sort: queryParams.sort,
       isPrerelease: latestRelease.prerelease,
     })
   }
